Refactored quicksave into multiple files for easier reading
