+++ quicksave/commands/__init__.py
@@ -1,0 +1,12 @@
+from .command_init import command_init
+from .command_register import command_register
+from .command_save import command_save
+from .command_revert import command_revert
+from .command_alias import command_alias
+from .command_list import command_list
+from .command_delete import command_delete
+from .command_lookup import command_lookup
+from .command_recover import command_recover
+from .command_clean import command_clean
+from .command_status import command_status
+from .command_config import command_config
